# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Time Tracking API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /leaves:
    get:
      operationId: listLeaves
      tags: [leaves]
      summary: List leave entries
      parameters:
        - name: startDate
          in: query
          schema:
            type: string
            format: date
        - name: endDate
          in: query
          schema:
            type: string
            format: date
        - name: userId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of leaves
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Leave"
    post:
      operationId: logLeave
      tags: [leaves]
      summary: Log a leave day
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LeaveInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Leave"
  /leaves/bulk:
    post:
      operationId: logLeavesBulk
      tags: [leaves]
      summary: Log multiple leave days at once (skips duplicates)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LeaveBulkInput"
      responses:
        "201":
          description: Bulk result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LeaveBulkResult"
  /leaves/{id}:
    delete:
      operationId: deleteLeave
      tags: [leaves]
      summary: Delete a leave entry
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"

  # ─── Reports (all authenticated roles; scope enforced server-side) ───────────
...
components:
...
  schemas:
...
    MessageResponse:
      type: object
      required: [message]
      properties:
        message:
          type: string

    # ─── Auth ────────────────────────────────────────────────────────────
...
    Leave:
      type: object
      required: [id, userId, userName, userRole, date, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: integer
        userName:
          type: string
        userRole:
          type: string
        userTitle:
          type: ["string", "null"]
          description: >-
            Overrides the rank label where the person's designation differs from it. Null shows the ordinary role label.
        date:
          type: string
          format: date
        portion:
          type: number
          description: "How much of the working day was taken: 1 for a full day, 0.5 for a half."
        note:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
    LeaveInput:
      type: object
      required: [date]
      properties:
        date:
          type: string
          format: date
        halfDay:
          type: boolean
          description: "Take half the working day rather than all of it. Defaults to false."
        note:
          type: string
    LeaveBulkInput:
      type: object
      required: [dates]
      properties:
        dates:
          type: array
          items:
            type: string
            format: date
          minItems: 1
        halfDay:
          type: boolean
          description: "Applies to every date in this request. Defaults to false."
        note:
          type: string
    LeaveBulkResult:
      type: object
      required: [created, skipped]
      properties:
        created:
          type: array
          items:
            $ref: "#/components/schemas/Leave"
        skipped:
          type: array
          description: Dates that already had a leave logged (duplicates ignored)
          items:
            type: string
            format: date