# lib/api-zod/src/generated/types/leave.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Time Tracking API
 * OpenAPI spec version: 0.1.0
 */

export interface Leave {
  id: number;
  userId: number;
  userName: string;
  userRole: string;
  /**
     * Overrides the rank label where the person's designation differs from it. Null shows the ordinary role label.
     * @nullable
     */
  userTitle?: string | null;
  date: Date;
  /** How much of the working day was taken: 1 for a full day, 0.5 for a half. */
  portion?: number;
  /** @nullable */
  note?: string | null;
  createdAt: Date;
}